+++ src/Common/src/System/Net/IPAddressParserStatics.cs
@@ -8,5 +8,6 @@ namespace System.Net
     {
         public const int IPv4AddressBytes = 4;
         public const int IPv6AddressBytes = 16;
+        public const int IPv6AddressShorts = IPv6AddressBytes / 2;
     }
 }
