+++ src/System.Net.Primitives/src/System/Net/IPv4AddressHelper.cs
@@ -165,7 +165,7 @@ namespace System
         {
             int numberBase = Decimal;
             char ch;
-            long[] parts = new long[4];
+            long* parts = stackalloc long[4];
             long currentValue = 0;
             bool atLeastOneChar = false;
