+++ src/System.Net.Primitives/tests/UnitTests/Fakes/IPv4AddressHelper.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System
+{
+    internal static class IPv4AddressHelper
+    {
+        internal const int Invalid = -1;
+        internal static unsafe long ParseNonCanonical(char* name, int start, ref int end, bool notImplicitFile) => 0;
+    }
+}
