+++ src/System.Net.Primitives/tests/UnitTests/System.Net.Primitives.UnitTests.Tests.csproj
@@ -66,9 +66,10 @@
 
   <ItemGroup>
     <Compile Include="CookieCollectionTest.cs" />
-    <Compile Include="Fakes\HostInformation.cs" />
-    <Compile Include="Fakes\IPAddressPal.cs" />
     <Compile Include="Fakes\CookieException.cs" />
+    <Compile Include="Fakes\HostInformation.cs" />
+    <Compile Include="Fakes\IPv4AddressHelper.cs" />
+    <Compile Include="Fakes\IPv6AddressHelper.cs" />
     <Compile Include="Fakes\Serialization.cs" Condition="'$(TargetGroup)' != 'netstandard' And '$(TargetGroup)' != 'netcoreapp'" />
     <Compile Include="$(CommonPath)\System\StringExtensions.cs">
       <Link>Common\System\StringExtensions.cs</Link>
